[build] all errors fixed, build pass locally
