+++ rvnllm/src/lib.rs
@@ -1,5 +1,2 @@
 pub mod parser;
-pub mod commands;
-
-pub use parser::load_model;
